backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
asyncpg==0.30.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
alembic==1.14.1
httpx==0.28.1
beautifulsoup4==4.13.3
playwright==1.50.0
pandas==2.2.3
numpy==2.2.2
scikit-learn==1.6.1
statsmodels==0.14.4
apscheduler==3.11.0
pytest==8.3.4
